[bt][hci] Add CommandChannel partial exclusivity
Some HCI commands will automatically fail if another command in a set is
currently running.  For example, the Remote Name Request command will
fail if Inquiry is currently running, and the Create Connection command
will also not run if Inquiry or Remote Name Request is running.

Add SendExclusiveCommand, which will wait for specific opcodes to be not
running before sending the command to the controller.  This manages the
mutual exclusion of different commands at the controller stage.

Refactor CommandChannel to track whether an event handler is
asynchronous within it's data, so that when an asynchronous one
finishes, we remove the pending command from the list of currently
running opcodes.  This also means asynchronous pending commands will remain
in the pending list until they are complete.

fx format-code too.

BT-284
BT-632

Test: HCI_CommandChannelTest.ExclusiveCommands

Change-Id: Ib7223a91f44ca8da63f49509dd2a51e489b5f6e0