[bt][l2cap] ERTM TxEngine: add ability to receive ACKs
When we receive a new acknowledgement, we should take the steps
required by the Process-ReqSeq action (see Core Spec v5.0, Volume 3,
Part A, Sec 8.6.5.6).

We can't take all of those steps yet, because we don't have the
necessary data structures (in particular, we're missing the
retransmission queue). But implement what we can with what we have
right now.

While there: improve documentation of an existing field.

Bug: BT-442
Test: bt-host-unittests

Change-Id: I1c966edc163442181785c063a0e361254f91b1da