[bt][bredr] Skip SDP discovery with no searches
When a BR/EDR connection completes, we may search for services on the
remote peer if any searches are registered.  If no searches are
required, we should not connect to do nothing.

Added a failure to FakeDomain when an expected outgoing L2CAP connection is
added but never connected to.

Also fx format-code

BT-861 #done
BT-857

Test: bt-host-unittests
      (updated GAP_BrEdrConnectionManagerTest.ServiceSearch,
       GAP_BrEdrConnectionManagerTest.AddServiceSearchAll,
       SDP_ServiceDiscovererTest.*)
Change-Id: Ib5730fa738718eba10ad29bd26695103cdfe01ee