[bt][l2cap] add a fuzzer for EnhancedRetransmissionModeRxEngine
Add a fuzzer for the code that processes L2CAP PDUs for a Channel
operating in Enhanced Retransmission Mode.

The design of this fuzzer is a bit more subtle than the fuzzer for
BasicModeRxEngine, because EnhancedRetransmissionModeRxEngine
maintains some state (the expected sequence number of the next inbound
frame, |next_seqnum_).

It's tempting to handle this by retaining a single Engine instance
across multiple calls to LLVMFuzzerTestOneInput(). But doing so would
make failures hard to reproduce. The problem being: calling into the
RxEngine with the errant input might behave differently depending on
whether the RxEngine is in a fresh state, or the RxEngine has
processed some earlier inputs.

Instead, we follow the advice of the libfuzzer documentation [*]. We
take the fuzz input data, and use it to generate a _sequence_ of
inputs to the RxEngine. Doing this with a fresh RxEngine on every call
to LLVMFuzzerTestOneInput() should ensure that failures are
reproducible.

[*] https://fuchsia.googlesource.com/fuchsia/+/master/docs/development/workflows/libfuzzer.md#Q_How-do-I-fuzz-more-complex-interfaces

Bug: BT-727
Test: manual (see below)

Manual test
-----------
$ fx set core.chromebook-x64 --fuzz-with asan --with //bundles:tests
$ fx clean
$ fx build
$ fx update  # or `fx push-package bluetooth_fuzzers` for incremental change
$ fx fuzz start enhanced_retransmission_mode_rx_engine_fuzzer

Change-Id: Ic147400d521ebd2c0711d118631c8e3363875dfc