[bt] Adapter now creates Data Domain after ACL initialization
The gap::Adapter now creates the data::Domain object, rather than
receiving it from Host. It creates it *after* the initialization of the
AclDataChannel, so that channel is guaranteed to live for the lifetime
of data::Domain.

data::Impl::CleanUp now checks that hci_->acl_data_channel() is non-null
before clearing the data handler. This should be the case because we
assure ordering, but this allows us to test that it is safe. CleanUp()
should be safe in all circumstances, and idempotent.

test:
* Added a unit test to ensure that we can safely shutdown a
  data::Domain that has not been initialized yet.
* Run bt-crate-unittests repeatedly in a loop and verify that the
  tests do not crash in data::Impl::CleanUp(), and instead
  continue to run successfully

Bug: 35072

Change-Id: I708df4b63dcebc4a17d3f1c3e7e0a831985efd8b