[bt][hci] Remove coalesced reports for advertising/scan response
hci::LegacyLowEnergyScanner doesn't immediately notify events from
scannable advertisements and waits until it receives a scan response
from the peer to provide a single coalesced report for peers found
during an active scan. In the case when a peer never sends a scan
response, the notification is deferred until the scan period gets
restarted.

This leads to issues in the field where devices often don't send a scan
response (particularly ADV_IND) which can cause a discovered device to
be reported with up to a 10 second delay (which is our scan period).

This logic has now been removed. Instead:
1. hci::LowEnergyScanner reports events as they arrive. Scan response
   events are annotated as such in the LowEnergyScanResult structure.
2. gap::LowEnergyDiscoveryManager has explicit handling for scan
   response events in which they are used to extend the advertising data
   structure of a previously found peer. This means that scan response
   events update their peer entries as soon as they arrive.

Bug: 35582
Test: bt-host-unittests
Change-Id: I2ae2a7d75e04481f1f83c2381f8523199c0a6e2b