[bt][gap-bredr] Wire PairingState to PairingDelegate actions
In PairingState, when a user interaction HCI event is received, call the
corresponding PairingDelegate action.

Move computation of expected events and security properties to a member
function of PairingState::Pairing.

Because PairingDelegate callbacks execute asynchronously and
out-of-thread, add a WeakPtrFactory to Pairing instances so that
pointers to a particular pairing procedure do not dangle nor alias to
each other.

Bug: 967

Test: In bt-host-unittests, GAP_PairingStateTest.*

Change-Id: Ib1e49ffa46867662e63414c038770e537461f78e