[bt] Return error for too-large NewSlabBuffer allocations
NewSlabBuffer, which selects the slabs from which blocks objects are
allocated, selects the "large" allocator for all requests too large for
the "small" allocator but does not check that the allocation is
successful nor does it propagate errors from the underlying allocators.

Add a check that the allocation size does not exceed the maximum
possible size for the "large" slab.

Additionally, upgrade the SlabBuffer size checks to include non-debug
builds. These are computationally inexpensive relative to the allocation
(expected to have ummeasurable effect on latency).

While there, add comments for NewSlabBuffer and add braces per "goto
fail" guidelines.

Bug: 38495

Test: in bt-host-unittests, SlabAllocatorTest.NewSlabBuffer
Change-Id: Ic9947240ba3148271c24deb2e6d5355ab3081601