[bt][host] Fix race condition in RunTaskSync()
Fix a race condition in RunTaskSync where it was possible for the
condition variable wait to be spuriously woken up *after* the done
condition was set but *before* the call to .notify(), such that the
condition variable did not live long enough and the call to .notify()
would crash.

I believe this is responsible for flakes in bt-integration-tests as seen
in fxb/49570.

Test: fx test fuchsia-pkg://fuchsia.com/bluetooth-tests#meta/bt-integration-tests.cmx

Bug: 49570
Change-Id: Ic4a8f27c3e8ecbef7435612bfac4a88d47341ea7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/385269
Commit-Queue: Nick Pollard <nickpollard@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Testability-Review: Arman Uguray <armansito@google.com>