[bt][hci-emulator] Fix use-after-free in LE connect
Fix a use-after-free bug (detected by ASAN) in the bluetooth
hci-emulator where we posted a delayed task that closed over a pointer
to a peer that is not guaranteed to live long enough. Instead close
over a weak pointer to the controller and the peer address and then use
those to re-acquire the peer reference (if it exists).

Bug: 49570
Test: bt-integration-tests

Not using a Multiply directive as there is currently at least one other
flake in bt-integration-tests that needs remedying, but this issue is
the more prominent one and needs an urgent fix.

Change-Id: Ib9a00041f8693a55e838f36e3de2a2d49c8a9adf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392233
Commit-Queue: Nick Pollard <nickpollard@google.com>
Testability-Review: Nick Pollard <nickpollard@google.com>
Reviewed-by: Xo Wang <xow@google.com>